requirements.txt
# Web
fastapi==0.115.2
uvicorn[standard]==0.30.6

# Validation
pydantic==2.8.2

# DB
SQLAlchemy==2.0.35
asyncpg==0.29.0
alembic==1.13.2

# Redis
redis==5.0.8

# LangChain & OpenAI（两者配套）
langchain==0.3.0
langchain-openai==0.2.4

# HTTP
httpx==0.27.2

# Env
python-dotenv==1.0.1

# Test
pytest==8.3.3
